Added basic notification service in ChatAPI layer
Added the notification service to use a basic concurrent queue.
A background worker thread continously monitors the queue for notifications. If any found, are pushed to clients
Once dequeued notifications are sent as blocking asynchronous code. So first notification is sent to all required clients before handling the 2nd notification. This needs improvement to make it truly asynchronous.